requirements.txt
requests
python-dotenv
schedule

